Integrate the reviewed Visual Perception and Region Discovery state from dev while preserving Feature Extraction contracts, artifacts, evaluation, and public exports.

Register dense feature resolution enhancement in DAG input validation and support fractional Region2D boxes through deterministic raster envelopes.

+++ src/contextmap/visual_perception/feature_diagnostics.py
@@ -393,7 +393,7 @@ def _encode_dense(item: DenseFeatureDiagnostic) -> dict[str, Any]:
     }
 
 
-def _encode_box(box: BoundingBox2D | None) -> dict[str, int] | None:
+def _encode_box(box: BoundingBox2D | None) -> dict[str, float] | None:
     if box is None:
         return None
     return {"x": box.x, "y": box.y, "width": box.width, "height": box.height}
